Complete the merging strategy for the controller
