--- src/controller/query/build.js
@@ -1,14 +1,0 @@
-// __Module Definition__
-module.exports = function() {
-  const controller = this;
-
-  controller.query('collection', '*', function(request, response, next) {
-    request.baucis.query = controller.model().find(request.baucis.conditions);
-    next();
-  });
-
-  controller.query('instance', '*', function(request, response, next) {
-    request.baucis.query = controller.model().findOne(request.baucis.conditions);
-    next();
-  });
-};
